# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SCIM (System for Cross-domain Identity Management) 2.0 实现，提供用户和组管理功能",
        "title": "SCIM 2.0 API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "shoy160",
            "email": "shoy160@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/scim/v2",
    "paths": {
...
        "/Users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户列表，支持分页、过滤和属性选择",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "列出所有用户",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "起始索引",
                        "name": "startIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "过滤条件",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "要返回的属性列表，逗号分隔",
                        "name": "attributes",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "要排除的属性列表，逗号分隔",
                        "name": "excludedAttributes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户列表",
                        "schema": {
                            "$ref": "#/definitions/model.ListResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新用户，需要提供用户名和姓名信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "用户已存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/Users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据用户ID获取用户详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "获取单个用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "要返回的属性列表，逗号分隔",
                        "name": "attributes",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "要排除的属性列表，逗号分隔",
                        "name": "excludedAttributes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户信息",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "全量替换用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "全量更新用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据用户ID删除用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功"
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "使用补丁操作更新用户属性",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "补丁更新用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "补丁操作",
                        "name": "patch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Email": {
            "type": "object",
            "properties": {
                "primary": {
                    "type": "boolean"
                },
                "type": {
                    "description": "work/home/other",
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detail 错误详情",
                    "type": "string"
                },
                "schemas": {
                    "description": "Schemas SCIM错误Schema",
                    "type": "string"
                },
                "scimType": {
                    "description": "ScimType SCIM错误类型",
                    "type": "string"
                },
                "status": {
                    "description": "Status HTTP状态码",
                    "type": "integer"
                }
            }
        },
...
        "model.ListResponse": {
            "type": "object",
            "properties": {
                "Resources": {
                    "description": "Resources 资源列表"
                },
                "itemsPerPage": {
                    "description": "ItemsPerPage 每页项目数",
                    "type": "integer"
                },
                "schemas": {
                    "description": "Schemas SCIM Schema标识符数组",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "startIndex": {
                    "description": "StartIndex 起始索引（从1开始）",
                    "type": "integer"
                },
                "totalResults": {
                    "description": "TotalResults 总结果数量",
                    "type": "integer"
                }
            }
        },
...
        "model.Meta": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "lastModified": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "resourceType": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "model.PatchRequest": {
            "type": "object",
            "required": [
                "Operations",
                "schemas"
            ],
            "properties": {
                "Operations": {
                    "description": "Operations 补丁操作列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PatchOperation"
                    }
                },
                "schemas": {
                    "description": "Schemas SCIM Schema标识符数组",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.Role": {
            "type": "object",
            "properties": {
                "display": {
                    "type": "string"
                },
                "primary": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "description": "角色值",
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "displayName": {
                    "type": "string"
                },
                "emails": {
                    "description": "邮箱（多值属性，关联表）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Email"
                    }
                },
                "externalId": {
                    "type": "string"
                },
                "groups": {
                    "description": "用户所属的组（非数据库字段，动态填充）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserGroup"
                    }
                },
                "id": {
                    "type": "string"
                },
                "meta": {
                    "description": "ResourceType由Meta.ResourceType动态生成，不持久化",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Meta"
                        }
                    ]
                },
                "name": {
                    "description": "姓名信息（嵌套）",
                    "type": "object",
                    "properties": {
                        "familyName": {
                            "type": "string"
                        },
                        "formatted": {
                            "type": "string"
                        },
                        "givenName": {
                            "type": "string"
                        },
                        "middleName": {
                            "type": "string"
                        }
                    }
                },
                "nickName": {
                    "type": "string"
                },
                "password": {
                    "description": "密码，非必填，默认不返回",
                    "type": "string"
                },
                "profileUrl": {
                    "type": "string"
                },
                "roles": {
                    "description": "角色（SCIM标准）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Role"
                    }
                },
                "schemas": {
                    "description": "存储SCIM schemas，包括自定义schemas",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "model.UserGroup": {
            "type": "object",
            "properties": {
                "display": {
                    "description": "组显示名称",
                    "type": "string"
                },
                "value": {
                    "description": "组ID",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "请输入 \"Bearer {token}\" 格式的认证令牌",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}